Refactoring the classes that work with the database.
Now the IDbManager and ISimplifiedDbManager developers create a new context when instantiating the class, it does not need to be passed from outside.

The OperationManager class is not officially static.

The DbContextOptions create method automatically detects where the appsettings.json file is located.

IAppController developers no longer have context properties.